Updated some images and folder names
