+++ masterhunter/masterhunter.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Master Hunter</title>
+    <script src="https://cdn.tailwindcss.com"></script>
+    <script src="./script.js" type="module" defer></script>
+</head>
+<body class="lex flex-col items-center bg-gray-800 text-white">
+    <header class="w-full bg-gray-900 p-4 text-2xl flex justify-between">
+        <div class="flex justify-center">
+            <h1 class="text-white">Monster Hunter Guessing Game</h1>
+        </div>
+        <div class="space-x-4 items-center">
+            <a href="../index.html" class="text-gray-300 hover:text-white">Guess the Monster</a>
+            <a href="./masterhunter.html" class="text-gray-300 hover:text-white underline underline-offset-2">Master Hunter</a>
+            <a href="#" class="text-gray-300 hover:text-white">Services</a>
+            <a href="#" class="text-gray-300 hover:text-white">About</a>
+        </div>
+    </header>
+    <main >
+        <div class="flex flex-1 flex-col items-center space-y-6">
+            <button onclick="fillresults()">Fill Results</button>
+            <h1 class="text-4xl font-bold">Name Every Monster!</h1>
+            <div id="search-bar-div">
+                <input id="search-bar" type="text" placeholder="Your Guess..." class="p-3 ps-10 rounded-lg bg-gray-600 focus:outline-none focus:ring-2 focus:ring-blue-500 justify-center align items-center" />
+                <div id="result" class="absolute left-1/2 -translate-x-1/2 grid grid-row border-0 border-white rounded-lg bg-white p-1 mt-2 z-20"></div>
+            </div>
+            <div>
+                <button onclick="sortArray()" class="border border-gray-600 rounded-full text-lg p-2 mx-2">Filter Results</button>
+                <button onclick="toggleGameMode()" class="border border-gray-600 rounded-full text-lg p-2 mx-2">Change Display</button>
+            </div>
+            <div class="sticky top-0">
+                <table class="table-fixed">
+                    <tr id="table" class="bg-gray-700 border border-gray-600"></tr>
+                    <tr id="monstercount"></tr>
+                </table>
+            </div>
+        </div>
+        <div class="w-full">
+            <div id="grid" class="flex flex-row flex-wrap gap-4 justify-center py-3 mx-[9%]"></div>
+        </div>
+    </main>
+</body>
+</html>
